// src/types/common.ts
export type Class = 1 | 2 | 3 | 4 | 5 | 6

export type StatLabel =
  | 'hp_flat'
  | 'hp_percent'
  | 'atk_flat'
  | 'atk_percent'
  | 'def_flat'
  | 'def_percent'
  | 'spd'
  | 'crate'
  | 'cdmg'
  | 'res'
  | 'acc'

export type SetLabel =
  | 'energy'
  | 'guard'
  | 'swift'
  | 'blade'
  | 'rage'
  | 'focus'
  | 'endure'
  | 'fatal'
  | 'despair'
  | 'vampire'
  | 'violent'
  | 'nemesis'
  | 'will'
  | 'shield'
  | 'revenge'
  | 'destroy'
  | 'fight'
  | 'determination'
  | 'enhance'
  | 'accuracy'
  | 'tolerance'